Update OAuth1aRequestFilter so that if security is not enabled we continue the filter chain.
Previously, the check for whether security was enabled or not was embedded deep in the authentication logic
and the fact that we didn't continue the filter chain was hard to spot. This commit refactors the logic so
that the check for whether security is enabled or not happens at the start of the method.