Redirect unauthenticated users to the login screen when viewing the AuthorizeConfirmationConfig screen.
Previously, unauthenticated users with read access to Jenkins would not be prompted to log in. This would cause an
exception and show the user a stack trace when we tried to store the token. As a workaround, we will redirect
unauthenticated users to the login screen if security is enabled.