Bugfix: fix anonymous user oauth handling (#219)
* Update OAuth1aRequestFilter so that if security is not enabled we continue the filter chain.

Previously, the check for whether security was enabled or not was embedded deep in the authentication logic
and the fact that we didn't continue the filter chain was hard to spot. This commit refactors the logic so
that the check for whether security is enabled or not happens at the start of the method.

* Redirect unauthenticated users to the login screen when viewing the AuthorizeConfirmationConfig screen.

Previously, unauthenticated users with read access to Jenkins would not be prompted to log in. This would cause an
exception and show the user a stack trace when we tried to store the token. As a workaround, we will redirect
unauthenticated users to the login screen if security is enabled.